requirements.txt
# Core Django and Database
Django>=4.2.0,<5.0.0
djongo>=1.3.6
pymongo>=4.0.0
dnspython>=2.2.1

# Django REST Framework
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0.0
drf-spectacular>=0.26.0

# Environment and Configuration
django-environ>=0.10.0
python-decouple>=3.8.0

# Authentication and Security
django-allauth>=0.54.0
cryptography>=40.0.0

# Utilities
Pillow>=9.5.0
python-dateutil>=2.8.2
pytz>=2023.3

# Development and Testing
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
factory-boy>=3.2.0
pytest>=7.3.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Production
gunicorn>=20.1.0
whitenoise>=6.4.0

# Monitoring and Logging
sentry-sdk>=1.25.0
celery>=5.2.0
redis>=4.5.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.3.0